+++ common/src/main/java/immersive_melodies/resources/Melody.java
@@ -3,7 +3,6 @@ package immersive_melodies.resources;
 import net.minecraft.network.PacketByteBuf;
 
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -12,7 +11,6 @@ public class Melody extends MelodyDescriptor {
     static final Melody DEFAULT = new Melody();
 
     private final List<Track> tracks = new LinkedList<>();
-    private Track primaryTrack;
 
     public Melody() {
         super("unknown");
@@ -46,14 +44,6 @@ public class Melody extends MelodyDescriptor {
 
     public void addTrack(Track track) {
         tracks.add(track);
-        primaryTrack = null;
-    }
-
-    public Track getPrimaryTrack() {
-        if (primaryTrack == null) {
-            primaryTrack = getTracks().stream().max(Comparator.comparingInt(m -> (int) (m.getNotes().size() * m.getNotes().stream().mapToInt(Note::getNote).distinct().count()))).orElse(DEFAULT.tracks.get(0));
-        }
-        return primaryTrack;
     }
 
     public void encode(PacketByteBuf b) {
@@ -63,5 +53,33 @@ public class Melody extends MelodyDescriptor {
         for (Track note : tracks) {
             note.encode(b);
         }
+    }
+
+    public void trim() {
+        int offset = getFirstNoteTime();
+        for (Track track : tracks) {
+            List<Note> newNotes = new LinkedList<>();
+            for (Note note : track.getNotes()) {
+                newNotes.add(new Note(
+                        note.getNote(),
+                        note.getVelocity(),
+                        note.getTime() - offset,
+                        note.getLength(),
+                        note.getSustain()
+                ));
+            }
+            track.setNotes(newNotes);
+        }
+    }
+
+    private int getFirstNoteTime() {
+        int earliestNote = Integer.MAX_VALUE;
+        for (Track track : tracks) {
+            List<Note> notes = track.getNotes();
+            if (!notes.isEmpty()) {
+                earliestNote = Math.min(earliestNote, notes.get(0).getTime());
+            }
+        }
+        return earliestNote;
     }
 }