+++ changelog.md
@@ -6,6 +6,9 @@
     * `uploadPermissionLevel` (2 for example, only allows OPs to upload)
     * `loadInbuiltMidis` To allow disabling the inbuilt midis
     * `mobInstrumentDropFactor` To adjust the chance of mobs dropping their instrument
+* Silence is now trimmed from the beginning of the midi
+* Removed automatic track selection to avoid confusion, tracks are now all enabled by default
+* Finally fixed de-syncs between tracks
 
 # 0.3.0